requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
SQLAlchemy==2.0.25
mysql-connector-python==8.3.0
Werkzeug==3.0.1
